modify scan endpoints to use rate limiter with extra traffic (#3992)

+++ apps/app/src/main/scala/com/daml/network/console/ScanAppReference.scala
@@ -166,7 +166,7 @@ abstract class ScanAppReference(
     }
 
   @Help.Summary("Get the available credit for a validator")
-  def getValidatorCredit(): Long =
+  def getValidatorCredit(): Double =
     consoleEnvironment.run {
       httpCommand(HttpScanAppClient.GetValidatorCredit())
     }
